third faz

+++ src/FilterByYear.java
@@ -6,7 +6,7 @@ public class FilterByYear extends Filter {
     }
 
     public boolean accept(Song song) {
-        int year = song.getReleaseYear();
+        Double year = song.getReleaseYear();
         for (int y : years) {
             if (year == y) {
                 return true;
